s/namespace/property_namespace/ to not break using libhal from C++. Pointed     out by Simone Gotti <simone.gotti@email.it> Patch from Kay Sievers <kay.sievers@vrfy.org> Here is support for recognition of hfs and hfsplus filesystems on a main     block device, like a CD. Only the first partition is returned with the     probe.

+++ hald/linux/volume_id/volume_id.h
@@ -60,7 +60,7 @@ struct volume_id {
 	unsigned char	*sbbuf;
 	unsigned int	sbbuf_len;
 	unsigned char	*seekbuf;
-	unsigned int	seekbuf_off;
+	unsigned long	seekbuf_off;
 	unsigned int	seekbuf_len;
 	int		fd_close;
 };
